src/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.5
django-ratelimit==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
httpie==2.2.0
idna==2.10
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
